requirements.txt
pypdf>=3.0.0
python-docx>=0.8.11
scikit-learn>=1.0.0
sentence-transformers>=2.0.0
torch>=1.9.0
numpy>=1.20.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
spacy>=3.7.2
